pyproject.toml
[project]
name = "slack-research-bot"
version = "0.1.0"
description = "Slack Research Bot"
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "aioconsole>=0.8.1",
    "slack-sdk>=3.35.0",
    "temporalio>=1.11.1",
    "vertexai>=1.71.1",
    "asyncio (>=3.4.3,<4.0.0)",
    "slack-bolt (>=1.23.0,<2.0.0)",
    "aiohttp (>=3.12.9,<4.0.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "openai (>=1.84.0,<2.0.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langchain-openai (>=0.3.21,<0.4.0)",
]
